빌드 타입 에러 해결
